Refactor: Consolidate hardware and ML utilities
This commit refactors the code to use existing utilities from `utils/` for hardware and ML common functionalities. This eliminates redundant code and promotes reuse.
