+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/IMPLEMENTATION_SUMMARY.md
@@ -6,7 +6,7 @@ I have successfully implemented comprehensive unified utilities that **leverage 
 
 ## 🔧 **What Was Implemented**
 
-### 1. **Enhanced ML Common Integration** (`ml_common_integration_enhanced.py`)
+### 1. **ML Common Integration** (`ml_common_integration.py`)
 - **Leverages existing utilities**: Uses proven tools from `utils/ml_commons/`
 - **Lookahead Protection**: Integrates `src.utils.ml_common.utils.lookahead_protection.LookaheadProtection`
 - **Overfitting Detection**: Uses `src.utils.ml_common.validation.enhanced_overfitting_detection.EnhancedOverfittingDetector`
@@ -33,10 +33,12 @@ I have successfully implemented comprehensive unified utilities that **leverage 
 - Experience replay and catastrophic forgetting prevention
 
 ### 5. **Optimized Hardware Management** (`unified_hardware_manager.py`)
+- **Leverages existing hardware utilities**: Uses `src.utils.hardware.unified_hardware_manager.UnifiedHardwareManager`
+- **Adaptive optimization**: Integrates `src.utils.hardware.adaptive_optimization_engine.AdaptiveOptimizationEngine`
+- **Advanced optimizers**: Uses existing CPU, GPU, and memory optimizers from `utils/hardware/`
 - GPU acceleration, memory optimization, and batch processing
 - Adaptive resource allocation based on workload type
 - Performance monitoring and optimization recommendations
-- Support for different optimization levels (basic to maximum)
 
 ### 6. **Unified Evaluation Framework** (`unified_evaluation_framework.py`)
 - Comprehensive evaluation including economic significance and trading viability
@@ -97,7 +99,7 @@ I have successfully implemented comprehensive unified utilities that **leverage 
 
 ## 🚀 **Usage Examples**
 
-### **Enhanced ML Common Integration**
+### **ML Common Integration**
 ```python
 from src.training.steps.market_analysis.hybrid_nas_tas_regime.shared_utils import (
     create_tas_ml_common_integration
@@ -147,15 +149,20 @@ tas_config.save_config("tas_config.json")
 ## 📁 **Files Created/Updated**
 
 ### **New Files**
-1. `ml_common_integration_enhanced.py` - Enhanced ML Common integration
+1. `ml_common_integration.py` - ML Common integration using existing utilities
 2. `unified_architecture_config.py` - Unified configuration management
 3. `unified_performance_monitor.py` - Performance monitoring system
 4. `unified_meta_learning.py` - Meta-learning utilities
-5. `unified_hardware_manager.py` - Hardware optimization
+5. `unified_hardware_manager.py` - Hardware optimization using existing utilities
 6. `unified_evaluation_framework.py` - Evaluation framework
 7. `examples/unified_utilities_example.py` - Working examples
 8. `UNIFIED_UTILITIES_INTEGRATION_GUIDE.md` - Integration guide
 9. `IMPLEMENTATION_SUMMARY.md` - This summary
+
+### **Deleted Files** (replaced by unified utilities)
+1. `hardware_optimization.py` - Replaced by unified_hardware_manager.py
+2. `unified_hardware_optimizer.py` - Replaced by unified_hardware_manager.py
+3. `ml_common_integration_enhanced.py` - Renamed to ml_common_integration.py
 
 ### **Updated Files**
 1. `__init__.py` - Updated imports and exports
