+++ src/training/steps/market_analysis/hybrid_nas_tas_regime/shared_utils/__init__.py
@@ -98,18 +98,11 @@ from .unified_validation_system import (
     create_unified_validation_system, quick_validation
 )
 
-# ML Common Integration - Enhanced integration with existing utilities
-from .ml_common_integration_enhanced import (
-    EnhancedMLCommonIntegration, MLCommonIntegrationType, MLCommonIntegrationConfig,
-    create_enhanced_ml_common_integration, create_tas_ml_common_integration,
-    create_nas_ml_common_integration, create_hybrid_ml_common_integration
-)
-
-# Legacy ML Common Integration (kept for backward compatibility)
+# ML Common Integration - Integration with existing utilities
 from .ml_common_integration import (
-    SharedMLUtilitiesManager, create_shared_ml_utilities_manager,
-    TASSharedMLUtilities, NASSharedMLUtilities, HybridSharedMLUtilities,
-    MLUtilityType, MLUtilityConfig
+    MLCommonIntegration, MLCommonIntegrationType, MLCommonIntegrationConfig,
+    create_ml_common_integration, create_tas_ml_common_integration,
+    create_nas_ml_common_integration, create_hybrid_ml_common_integration
 )
 
 # New Unified Architecture Management
@@ -211,15 +204,10 @@ __all__ = [
     'create_unified_config_manager', 'load_config_from_file', 'create_environment_config',
     'create_unified_validation_system', 'quick_validation',
 
-    # ML Common Integration - Enhanced integration with existing utilities
-    'EnhancedMLCommonIntegration', 'MLCommonIntegrationType', 'MLCommonIntegrationConfig',
-    'create_enhanced_ml_common_integration', 'create_tas_ml_common_integration',
+    # ML Common Integration - Integration with existing utilities
+    'MLCommonIntegration', 'MLCommonIntegrationType', 'MLCommonIntegrationConfig',
+    'create_ml_common_integration', 'create_tas_ml_common_integration',
     'create_nas_ml_common_integration', 'create_hybrid_ml_common_integration',
-
-    # Legacy ML Common Integration (kept for backward compatibility)
-    'SharedMLUtilitiesManager', 'create_shared_ml_utilities_manager',
-    'TASSharedMLUtilities', 'NASSharedMLUtilities', 'HybridSharedMLUtilities',
-    'MLUtilityType', 'MLUtilityConfig',
 
     # New Unified Architecture Management
     'BaseArchitectureConfig', 'TASArchitectureConfig', 'NASArchitectureConfig', 'HybridArchitectureConfig',
