--- src/training/steps/market_analysis/hybrid_nas_tas_regime/shared_utils/unified_hardware_optimizer.py
@@ -1,1105 +1,0 @@
-"""
-Unified Hardware Optimizer
-
-This module provides a unified hardware optimization system that combines
-the best practices from both TAS and NAS regime detection systems. It provides
-comprehensive hardware acceleration and optimization for regime detection systems.
-
-Features:
-- Unified hardware optimization for both tree and neural architectures
-- Support for multiple hardware backends (CPU, GPU, specialized accelerators)
-- Memory optimization and management
-- Performance monitoring and profiling
-- Adaptive optimization based on workload characteristics
-- Real-time optimization capabilities
-"""
-
-import numpy as np
-import pandas as pd
-from typing import Dict, List, Any, Optional, Tuple, Union, Callable
-import logging
-from dataclasses import dataclass, field
-from enum import Enum
-import time
-from datetime import datetime
-from contextlib import contextmanager
-import warnings
-warnings.filterwarnings('ignore')
-
-logger = logging.getLogger(__name__)
-
-
-class HardwareBackend(Enum):
-    """Types of hardware backends."""
-    CPU = "cpu"
-    GPU = "gpu"
-    M1_GPU = "m1_gpu"
-    SPECIALIZED = "specialized"
-
-
-class OptimizationLevel(Enum):
-    """Hardware optimization levels."""
-    NONE = "none"
-    BASIC = "basic"
-    AGGRESSIVE = "aggressive"
-    MAXIMUM = "maximum"
-
-
-class WorkloadType(Enum):
-    """Types of workloads."""
-    REGIME_DETECTION = "regime_detection"
-    TREE_OPERATIONS = "tree_operations"
-    NEURAL_INFERENCE = "neural_inference"
-    CLUSTERING = "clustering"
-    STATISTICAL_ANALYSIS = "statistical_analysis"
-    MATRIX_OPERATIONS = "matrix_operations"
-
-
-@dataclass
-class HardwareConfig:
-    """Configuration for unified hardware optimization."""
-    
-    # Hardware backends
-    primary_backend: HardwareBackend = HardwareBackend.CPU
-    enable_gpu: bool = True
-    enable_m1_gpu: bool = True
-    enable_specialized: bool = False
-    
-    # Optimization levels
-    cpu_optimization_level: OptimizationLevel = OptimizationLevel.AGGRESSIVE
-    gpu_optimization_level: OptimizationLevel = OptimizationLevel.AGGRESSIVE
-    memory_optimization_level: OptimizationLevel = OptimizationLevel.AGGRESSIVE
-    
-    # Memory management
-    max_memory_usage_gb: float = 8.0
-    enable_memory_cleanup: bool = True
-    memory_cleanup_interval: int = 1000
-    enable_memory_mapping: bool = True
-    
-    # Performance monitoring
-    enable_performance_monitoring: bool = True
-    monitoring_interval: int = 100
-    enable_profiling: bool = False
-    enable_benchmarking: bool = True
-    
-    # Adaptive optimization
-    enable_adaptive_optimization: bool = True
-    adaptation_threshold: float = 0.1  # 10% performance improvement threshold
-    enable_learning: bool = True
-    learning_rate: float = 0.01
-    
-    # Parallel processing
-    enable_parallel_processing: bool = True
-    n_workers: int = 4
-    parallel_evaluation: bool = True
-    
-    # Workload-specific optimizations
-    enable_regime_optimization: bool = True
-    enable_tree_optimization: bool = True
-    enable_neural_optimization: bool = True
-    enable_clustering_optimization: bool = True
-    enable_statistical_optimization: bool = True
-
-
-@dataclass
-class PerformanceMetrics:
-    """Hardware performance metrics."""
-    
-    # Execution metrics
-    execution_time: float = 0.0
-    throughput: float = 0.0
-    latency: float = 0.0
-    
-    # Memory metrics
-    memory_usage_mb: float = 0.0
-    memory_peak_mb: float = 0.0
-    memory_efficiency: float = 0.0
-    
-    # Hardware utilization
-    cpu_utilization: float = 0.0
-    gpu_utilization: float = 0.0
-    memory_utilization: float = 0.0
-    
-    # Optimization metrics
-    optimization_gain: float = 0.0
-    cache_hit_rate: float = 0.0
-    parallel_efficiency: float = 0.0
-    
-    # Workload-specific metrics
-    regime_detection_time: float = 0.0
-    tree_operation_time: float = 0.0
-    neural_inference_time: float = 0.0
-    clustering_time: float = 0.0
-    statistical_analysis_time: float = 0.0
-
-
-class UnifiedHardwareOptimizer:
-    """
-    Unified Hardware Optimizer.
-    
-    Combines the best practices from both TAS and NAS regime detection systems
-    to provide comprehensive hardware optimization.
-    """
-    
-    def __init__(self, config: HardwareConfig):
-        """Initialize unified hardware optimizer.
-        
-        Args:
-            config: Hardware optimization configuration
-        """
-        self.config = config
-        self.logger = logging.getLogger(self.__class__.__name__)
-        
-        # Initialize hardware components
-        self.hardware_components = {}
-        self.performance_monitor = None
-        self.memory_manager = None
-        
-        # Performance tracking
-        self.performance_history = []
-        self.optimization_history = []
-        self.workload_profiles = {}
-        
-        # Initialize components
-        self._initialize_hardware_components()
-        self._initialize_performance_monitoring()
-        self._initialize_memory_management()
-        
-        self.logger.info("✅ Unified Hardware Optimizer initialized")
-        self._log_initialization_status()
-    
-    def _initialize_hardware_components(self):
-        """Initialize hardware optimization components."""
-        try:
-            # Initialize CPU optimization
-            if self.config.cpu_optimization_level != OptimizationLevel.NONE:
-                self.hardware_components['cpu'] = self._create_cpu_optimizer()
-            
-            # Initialize GPU optimization
-            if self.config.enable_gpu and self.config.gpu_optimization_level != OptimizationLevel.NONE:
-                self.hardware_components['gpu'] = self._create_gpu_optimizer()
-            
-            # Initialize M1 GPU optimization
-            if self.config.enable_m1_gpu:
-                self.hardware_components['m1_gpu'] = self._create_m1_gpu_optimizer()
-            
-            # Initialize specialized accelerators
-            if self.config.enable_specialized:
-                self.hardware_components['specialized'] = self._create_specialized_optimizer()
-            
-            self.logger.info("✅ Hardware components initialized")
-            
-        except Exception as e:
-            self.logger.error(f"Failed to initialize hardware components: {e}")
-    
-    def _create_cpu_optimizer(self) -> Dict[str, Any]:
-        """Create CPU optimizer."""
-        return {
-            'type': 'cpu',
-            'optimization_level': self.config.cpu_optimization_level,
-            'n_workers': self.config.n_workers,
-            'enable_parallel': self.config.enable_parallel_processing,
-            'enable_vectorization': True,
-            'enable_optimization_flags': True
-        }
-    
-    def _create_gpu_optimizer(self) -> Dict[str, Any]:
-        """Create GPU optimizer."""
-        return {
-            'type': 'gpu',
-            'optimization_level': self.config.gpu_optimization_level,
-            'enable_cuda': True,
-            'enable_tensor_cores': True,
-            'enable_mixed_precision': True,
-            'memory_fraction': 0.8
-        }
-    
-    def _create_m1_gpu_optimizer(self) -> Dict[str, Any]:
-        """Create M1 GPU optimizer."""
-        return {
-            'type': 'm1_gpu',
-            'optimization_level': self.config.gpu_optimization_level,
-            'enable_metal': True,
-            'enable_neural_engine': True,
-            'enable_shared_memory': True
-        }
-    
-    def _create_specialized_optimizer(self) -> Dict[str, Any]:
-        """Create specialized accelerator optimizer."""
-        return {
-            'type': 'specialized',
-            'optimization_level': OptimizationLevel.AGGRESSIVE,
-            'enable_tpu': False,  # Would be True if TPU available
-            'enable_fpga': False,  # Would be True if FPGA available
-            'enable_asic': False   # Would be True if ASIC available
-        }
-    
-    def _initialize_performance_monitoring(self):
-        """Initialize performance monitoring."""
-        if not self.config.enable_performance_monitoring:
-            return
-        
-        try:
-            self.performance_monitor = {
-                'enabled': True,
-                'monitoring_interval': self.config.monitoring_interval,
-                'enable_profiling': self.config.enable_profiling,
-                'enable_benchmarking': self.config.enable_benchmarking,
-                'metrics': PerformanceMetrics()
-            }
-            
-            self.logger.info("✅ Performance monitoring initialized")
-            
-        except Exception as e:
-            self.logger.error(f"Failed to initialize performance monitoring: {e}")
-            self.performance_monitor = None
-    
-    def _initialize_memory_management(self):
-        """Initialize memory management."""
-        if not self.config.enable_memory_cleanup:
-            return
-        
-        try:
-            self.memory_manager = {
-                'enabled': True,
-                'max_memory_gb': self.config.max_memory_usage_gb,
-                'cleanup_interval': self.config.memory_cleanup_interval,
-                'enable_mapping': self.config.enable_memory_mapping,
-                'current_usage_mb': 0.0,
-                'peak_usage_mb': 0.0
-            }
-            
-            self.logger.info("✅ Memory management initialized")
-            
-        except Exception as e:
-            self.logger.error(f"Failed to initialize memory management: {e}")
-            self.memory_manager = None
-    
-    def _log_initialization_status(self):
-        """Log the initialization status of all components."""
-        self.logger.info("🔧 Hardware Optimization Status:")
-        self.logger.info(f"   CPU Optimizer: {'✅' if 'cpu' in self.hardware_components else '❌'}")
-        self.logger.info(f"   GPU Optimizer: {'✅' if 'gpu' in self.hardware_components else '❌'}")
-        self.logger.info(f"   M1 GPU Optimizer: {'✅' if 'm1_gpu' in self.hardware_components else '❌'}")
-        self.logger.info(f"   Specialized Optimizer: {'✅' if 'specialized' in self.hardware_components else '❌'}")
-        self.logger.info(f"   Performance Monitor: {'✅' if self.performance_monitor else '❌'}")
-        self.logger.info(f"   Memory Manager: {'✅' if self.memory_manager else '❌'}")
-    
-    @contextmanager
-    def optimization_context(self, workload_type: WorkloadType):
-        """Context manager for hardware optimization."""
-        try:
-            # Start optimization
-            self._start_optimization(workload_type)
-            yield
-        finally:
-            # Stop optimization
-            self._stop_optimization(workload_type)
-    
-    def _start_optimization(self, workload_type: WorkloadType):
-        """Start hardware optimization for workload."""
-        try:
-            # Configure hardware for workload
-            self._configure_hardware_for_workload(workload_type)
-            
-            # Start performance monitoring
-            if self.performance_monitor:
-                self._start_performance_monitoring(workload_type)
-            
-            # Start memory management
-            if self.memory_manager:
-                self._start_memory_management()
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to start optimization: {e}")
-    
-    def _stop_optimization(self, workload_type: WorkloadType):
-        """Stop hardware optimization for workload."""
-        try:
-            # Stop performance monitoring
-            if self.performance_monitor:
-                self._stop_performance_monitoring(workload_type)
-            
-            # Cleanup memory if needed
-            if self.memory_manager:
-                self._cleanup_memory_if_needed()
-            
-            # Record optimization results
-            self._record_optimization_results(workload_type)
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to stop optimization: {e}")
-    
-    def _configure_hardware_for_workload(self, workload_type: WorkloadType):
-        """Configure hardware for specific workload type."""
-        try:
-            if workload_type == WorkloadType.REGIME_DETECTION:
-                self._configure_for_regime_detection()
-            elif workload_type == WorkloadType.TREE_OPERATIONS:
-                self._configure_for_tree_operations()
-            elif workload_type == WorkloadType.NEURAL_INFERENCE:
-                self._configure_for_neural_inference()
-            elif workload_type == WorkloadType.CLUSTERING:
-                self._configure_for_clustering()
-            elif workload_type == WorkloadType.STATISTICAL_ANALYSIS:
-                self._configure_for_statistical_analysis()
-            elif workload_type == WorkloadType.MATRIX_OPERATIONS:
-                self._configure_for_matrix_operations()
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to configure hardware for {workload_type.value}: {e}")
-    
-    def _configure_for_regime_detection(self):
-        """Configure hardware for regime detection workload."""
-        # Optimize for regime detection specific operations
-        if 'cpu' in self.hardware_components:
-            self.hardware_components['cpu']['enable_vectorization'] = True
-            self.hardware_components['cpu']['enable_parallel'] = True
-        
-        if 'gpu' in self.hardware_components:
-            self.hardware_components['gpu']['enable_mixed_precision'] = True
-            self.hardware_components['gpu']['memory_fraction'] = 0.9
-    
-    def _configure_for_tree_operations(self):
-        """Configure hardware for tree operations workload."""
-        # Optimize for tree-based operations
-        if 'cpu' in self.hardware_components:
-            self.hardware_components['cpu']['enable_parallel'] = True
-            self.hardware_components['cpu']['n_workers'] = min(self.config.n_workers, 8)
-    
-    def _configure_for_neural_inference(self):
-        """Configure hardware for neural inference workload."""
-        # Optimize for neural network operations
-        if 'gpu' in self.hardware_components:
-            self.hardware_components['gpu']['enable_tensor_cores'] = True
-            self.hardware_components['gpu']['enable_mixed_precision'] = True
-        
-        if 'm1_gpu' in self.hardware_components:
-            self.hardware_components['m1_gpu']['enable_neural_engine'] = True
-    
-    def _configure_for_clustering(self):
-        """Configure hardware for clustering workload."""
-        # Optimize for clustering algorithms
-        if 'cpu' in self.hardware_components:
-            self.hardware_components['cpu']['enable_parallel'] = True
-            self.hardware_components['cpu']['enable_vectorization'] = True
-    
-    def _configure_for_statistical_analysis(self):
-        """Configure hardware for statistical analysis workload."""
-        # Optimize for statistical computations
-        if 'cpu' in self.hardware_components:
-            self.hardware_components['cpu']['enable_parallel'] = True
-            self.hardware_components['cpu']['enable_optimization_flags'] = True
-    
-    def _configure_for_matrix_operations(self):
-        """Configure hardware for matrix operations workload."""
-        # Optimize for matrix operations
-        if 'gpu' in self.hardware_components:
-            self.hardware_components['gpu']['enable_tensor_cores'] = True
-            self.hardware_components['gpu']['memory_fraction'] = 0.8
-    
-    def _start_performance_monitoring(self, workload_type: WorkloadType):
-        """Start performance monitoring."""
-        try:
-            if self.performance_monitor:
-                self.performance_monitor['start_time'] = time.time()
-                self.performance_monitor['workload_type'] = workload_type.value
-                self.performance_monitor['metrics'] = PerformanceMetrics()
-                
-                # Start system monitoring
-                self._start_system_monitoring()
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to start performance monitoring: {e}")
-    
-    def _stop_performance_monitoring(self, workload_type: WorkloadType):
-        """Stop performance monitoring."""
-        try:
-            if self.performance_monitor:
-                end_time = time.time()
-                start_time = self.performance_monitor.get('start_time', end_time)
-                
-                # Calculate execution time
-                execution_time = end_time - start_time
-                self.performance_monitor['metrics'].execution_time = execution_time
-                
-                # Calculate throughput
-                if execution_time > 0:
-                    # This would be calculated based on actual workload
-                    throughput = 1000 / execution_time  # Placeholder
-                    self.performance_monitor['metrics'].throughput = throughput
-                
-                # Record performance metrics
-                self._record_performance_metrics(workload_type, execution_time)
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to stop performance monitoring: {e}")
-    
-    def _start_system_monitoring(self):
-        """Start system resource monitoring."""
-        try:
-            import psutil
-            
-            # Get initial system state
-            cpu_percent = psutil.cpu_percent()
-            memory = psutil.virtual_memory()
-            
-            if self.performance_monitor:
-                self.performance_monitor['metrics'].cpu_utilization = cpu_percent
-                self.performance_monitor['metrics'].memory_utilization = memory.percent
-                self.performance_monitor['metrics'].memory_usage_mb = memory.used / (1024 * 1024)
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to start system monitoring: {e}")
-    
-    def _start_memory_management(self):
-        """Start memory management."""
-        try:
-            if self.memory_manager:
-                import psutil
-                memory = psutil.virtual_memory()
-                
-                self.memory_manager['current_usage_mb'] = memory.used / (1024 * 1024)
-                self.memory_manager['peak_usage_mb'] = memory.used / (1024 * 1024)
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to start memory management: {e}")
-    
-    def _cleanup_memory_if_needed(self):
-        """Cleanup memory if usage exceeds threshold."""
-        try:
-            if not self.memory_manager:
-                return
-            
-            import psutil
-            memory = psutil.virtual_memory()
-            current_usage_gb = memory.used / (1024 * 1024 * 1024)
-            
-            if current_usage_gb > self.config.max_memory_usage_gb:
-                self.logger.warning(f"Memory usage ({current_usage_gb:.2f} GB) exceeds threshold ({self.config.max_memory_usage_gb} GB)")
-                
-                # Force garbage collection
-                import gc
-                gc.collect()
-                
-                # Update memory usage
-                memory = psutil.virtual_memory()
-                self.memory_manager['current_usage_mb'] = memory.used / (1024 * 1024)
-                
-        except Exception as e:
-            self.logger.warning(f"Failed to cleanup memory: {e}")
-    
-    def _record_performance_metrics(self, workload_type: WorkloadType, execution_time: float):
-        """Record performance metrics."""
-        try:
-            if not self.performance_monitor:
-                return
-            
-            metrics = self.performance_monitor['metrics']
-            
-            # Record workload-specific metrics
-            if workload_type == WorkloadType.REGIME_DETECTION:
-                metrics.regime_detection_time = execution_time
-            elif workload_type == WorkloadType.TREE_OPERATIONS:
-                metrics.tree_operation_time = execution_time
-            elif workload_type == WorkloadType.NEURAL_INFERENCE:
-                metrics.neural_inference_time = execution_time
-            elif workload_type == WorkloadType.CLUSTERING:
-                metrics.clustering_time = execution_time
-            elif workload_type == WorkloadType.STATISTICAL_ANALYSIS:
-                metrics.statistical_analysis_time = execution_time
-            
-            # Add to performance history
-            self.performance_history.append({
-                'timestamp': datetime.now(),
-                'workload_type': workload_type.value,
-                'execution_time': execution_time,
-                'memory_usage_mb': metrics.memory_usage_mb,
-                'cpu_utilization': metrics.cpu_utilization,
-                'gpu_utilization': metrics.gpu_utilization
-            })
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to record performance metrics: {e}")
-    
-    def _record_optimization_results(self, workload_type: WorkloadType):
-        """Record optimization results."""
-        try:
-            if not self.performance_monitor:
-                return
-            
-            metrics = self.performance_monitor['metrics']
-            
-            # Calculate optimization gain
-            optimization_gain = self._calculate_optimization_gain(workload_type)
-            metrics.optimization_gain = optimization_gain
-            
-            # Record optimization history
-            self.optimization_history.append({
-                'timestamp': datetime.now(),
-                'workload_type': workload_type.value,
-                'optimization_gain': optimization_gain,
-                'execution_time': metrics.execution_time,
-                'memory_usage_mb': metrics.memory_usage_mb
-            })
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to record optimization results: {e}")
-    
-    def _calculate_optimization_gain(self, workload_type: WorkloadType) -> float:
-        """Calculate optimization gain for workload."""
-        try:
-            # This would compare with baseline performance
-            # For now, return a placeholder value
-            return np.random.uniform(0.1, 0.3)  # 10-30% improvement
-            
-        except Exception as e:
-            self.logger.warning(f"Failed to calculate optimization gain: {e}")
-            return 0.0
-    
-    def optimize_regime_detection(self, 
-                                data: np.ndarray,
-                                regime_config: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Optimize regime detection operations.
-        
-        Args:
-            data: Input data for regime detection
-            regime_config: Regime detection configuration
-            
-        Returns:
-            Optimized regime detection results
-        """
-        try:
-            start_time = time.time()
-            
-            with self.optimization_context(WorkloadType.REGIME_DETECTION):
-                # Apply hardware optimizations
-                optimized_data = self._optimize_data_for_regime_detection(data)
-                
-                # Perform regime detection (simplified)
-                result = self._perform_optimized_regime_detection(optimized_data, regime_config)
-                
-                # Apply post-processing optimizations
-                result = self._optimize_regime_detection_results(result)
-            
-            execution_time = time.time() - start_time
-            self._track_performance('regime_detection', execution_time, len(data))
-            
-            return result
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize regime detection: {e}")
-            return self._fallback_regime_detection(data, regime_config)
-    
-    def optimize_tree_operations(self, 
-                               data: np.ndarray,
-                               tree_config: Dict[str, Any]) -> np.ndarray:
-        """
-        Optimize tree-based operations.
-        
-        Args:
-            data: Input data for tree operations
-            tree_config: Tree configuration parameters
-            
-        Returns:
-            Optimized tree operation results
-        """
-        try:
-            start_time = time.time()
-            
-            with self.optimization_context(WorkloadType.TREE_OPERATIONS):
-                # Apply hardware optimizations
-                optimized_data = self._optimize_data_for_tree_operations(data)
-                
-                # Perform tree operations (simplified)
-                result = self._perform_optimized_tree_operations(optimized_data, tree_config)
-            
-            execution_time = time.time() - start_time
-            self._track_performance('tree_operations', execution_time, len(data))
-            
-            return result
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize tree operations: {e}")
-            return self._fallback_tree_operations(data, tree_config)
-    
-    def optimize_neural_inference(self, 
-                                data: np.ndarray,
-                                neural_config: Dict[str, Any]) -> np.ndarray:
-        """
-        Optimize neural network inference.
-        
-        Args:
-            data: Input data for neural inference
-            neural_config: Neural network configuration
-            
-        Returns:
-            Optimized neural inference results
-        """
-        try:
-            start_time = time.time()
-            
-            with self.optimization_context(WorkloadType.NEURAL_INFERENCE):
-                # Apply hardware optimizations
-                optimized_data = self._optimize_data_for_neural_inference(data)
-                
-                # Perform neural inference (simplified)
-                result = self._perform_optimized_neural_inference(optimized_data, neural_config)
-            
-            execution_time = time.time() - start_time
-            self._track_performance('neural_inference', execution_time, len(data))
-            
-            return result
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize neural inference: {e}")
-            return self._fallback_neural_inference(data, neural_config)
-    
-    def optimize_clustering(self, 
-                          data: np.ndarray,
-                          clustering_config: Dict[str, Any]) -> np.ndarray:
-        """
-        Optimize clustering operations.
-        
-        Args:
-            data: Input data for clustering
-            clustering_config: Clustering configuration parameters
-            
-        Returns:
-            Optimized clustering results
-        """
-        try:
-            start_time = time.time()
-            
-            with self.optimization_context(WorkloadType.CLUSTERING):
-                # Apply hardware optimizations
-                optimized_data = self._optimize_data_for_clustering(data)
-                
-                # Perform clustering (simplified)
-                result = self._perform_optimized_clustering(optimized_data, clustering_config)
-            
-            execution_time = time.time() - start_time
-            self._track_performance('clustering', execution_time, len(data))
-            
-            return result
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize clustering: {e}")
-            return self._fallback_clustering(data, clustering_config)
-    
-    def optimize_statistical_analysis(self, 
-                                   data: np.ndarray,
-                                   statistical_config: Dict[str, Any]) -> Dict[str, Any]:
-        """
-        Optimize statistical analysis operations.
-        
-        Args:
-            data: Input data for statistical analysis
-            statistical_config: Statistical analysis configuration
-            
-        Returns:
-            Optimized statistical analysis results
-        """
-        try:
-            start_time = time.time()
-            
-            with self.optimization_context(WorkloadType.STATISTICAL_ANALYSIS):
-                # Apply hardware optimizations
-                optimized_data = self._optimize_data_for_statistical_analysis(data)
-                
-                # Perform statistical analysis (simplified)
-                result = self._perform_optimized_statistical_analysis(optimized_data, statistical_config)
-            
-            execution_time = time.time() - start_time
-            self._track_performance('statistical_analysis', execution_time, len(data))
-            
-            return result
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize statistical analysis: {e}")
-            return self._fallback_statistical_analysis(data, statistical_config)
-    
-    def _optimize_data_for_regime_detection(self, data: np.ndarray) -> np.ndarray:
-        """Optimize data for regime detection."""
-        try:
-            # Apply data preprocessing optimizations
-            if self.memory_manager and self.memory_manager['enabled']:
-                # Optimize memory layout
-                data = np.ascontiguousarray(data)
-            
-            # Apply CPU optimizations
-            if 'cpu' in self.hardware_components:
-                # Vectorization optimizations
-                data = self._apply_vectorization_optimizations(data)
-            
-            # Apply GPU optimizations
-            if 'gpu' in self.hardware_components:
-                # GPU memory optimizations
-                data = self._apply_gpu_memory_optimizations(data)
-            
-            return data
-            
-        except Exception as e:
-            self.logger.warning(f"Data optimization failed: {e}")
-            return data
-    
-    def _optimize_data_for_tree_operations(self, data: np.ndarray) -> np.ndarray:
-        """Optimize data for tree operations."""
-        try:
-            # Apply tree-specific optimizations
-            if 'cpu' in self.hardware_components:
-                # Parallel processing optimizations
-                data = self._apply_parallel_optimizations(data)
-            
-            return data
-            
-        except Exception as e:
-            self.logger.warning(f"Tree data optimization failed: {e}")
-            return data
-    
-    def _optimize_data_for_neural_inference(self, data: np.ndarray) -> np.ndarray:
-        """Optimize data for neural inference."""
-        try:
-            # Apply neural-specific optimizations
-            if 'gpu' in self.hardware_components:
-                # GPU tensor optimizations
-                data = self._apply_tensor_optimizations(data)
-            
-            if 'm1_gpu' in self.hardware_components:
-                # M1 GPU optimizations
-                data = self._apply_m1_gpu_optimizations(data)
-            
-            return data
-            
-        except Exception as e:
-            self.logger.warning(f"Neural data optimization failed: {e}")
-            return data
-    
-    def _optimize_data_for_clustering(self, data: np.ndarray) -> np.ndarray:
-        """Optimize data for clustering."""
-        try:
-            # Apply clustering-specific optimizations
-            if 'cpu' in self.hardware_components:
-                # Distance calculation optimizations
-                data = self._apply_distance_optimizations(data)
-            
-            return data
-            
-        except Exception as e:
-            self.logger.warning(f"Clustering data optimization failed: {e}")
-            return data
-    
-    def _optimize_data_for_statistical_analysis(self, data: np.ndarray) -> np.ndarray:
-        """Optimize data for statistical analysis."""
-        try:
-            # Apply statistical-specific optimizations
-            if 'cpu' in self.hardware_components:
-                # Statistical computation optimizations
-                data = self._apply_statistical_optimizations(data)
-            
-            return data
-            
-        except Exception as e:
-            self.logger.warning(f"Statistical data optimization failed: {e}")
-            return data
-    
-    def _apply_vectorization_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply vectorization optimizations."""
-        # Placeholder for vectorization optimizations
-        return data
-    
-    def _apply_gpu_memory_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply GPU memory optimizations."""
-        # Placeholder for GPU memory optimizations
-        return data
-    
-    def _apply_parallel_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply parallel processing optimizations."""
-        # Placeholder for parallel optimizations
-        return data
-    
-    def _apply_tensor_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply tensor optimizations."""
-        # Placeholder for tensor optimizations
-        return data
-    
-    def _apply_m1_gpu_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply M1 GPU optimizations."""
-        # Placeholder for M1 GPU optimizations
-        return data
-    
-    def _apply_distance_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply distance calculation optimizations."""
-        # Placeholder for distance optimizations
-        return data
-    
-    def _apply_statistical_optimizations(self, data: np.ndarray) -> np.ndarray:
-        """Apply statistical computation optimizations."""
-        # Placeholder for statistical optimizations
-        return data
-    
-    def _perform_optimized_regime_detection(self, data: np.ndarray, config: Dict[str, Any]) -> Dict[str, Any]:
-        """Perform optimized regime detection."""
-        try:
-            # Simplified regime detection
-            n_regimes = config.get('n_regimes', 3)
-            regime_predictions = np.random.randint(0, n_regimes, len(data))
-            regime_probabilities = np.random.rand(len(data), n_regimes)
-            
-            return {
-                'regime_predictions': regime_predictions,
-                'regime_probabilities': regime_probabilities,
-                'regime_stability': np.random.rand(len(data)),
-                'transition_probabilities': np.random.rand(n_regimes, n_regimes)
-            }
-            
-        except Exception as e:
-            self.logger.error(f"Optimized regime detection failed: {e}")
-            return {}
-    
-    def _perform_optimized_tree_operations(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Perform optimized tree operations."""
-        try:
-            # Simplified tree operations
-            n_features = config.get('n_features', 10)
-            return np.random.rand(len(data), n_features)
-            
-        except Exception as e:
-            self.logger.error(f"Optimized tree operations failed: {e}")
-            return np.zeros((len(data), config.get('n_features', 10)))
-    
-    def _perform_optimized_neural_inference(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Perform optimized neural inference."""
-        try:
-            # Simplified neural inference
-            output_dim = config.get('output_dim', 1)
-            return np.random.rand(len(data), output_dim)
-            
-        except Exception as e:
-            self.logger.error(f"Optimized neural inference failed: {e}")
-            return np.zeros((len(data), config.get('output_dim', 1)))
-    
-    def _perform_optimized_clustering(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Perform optimized clustering."""
-        try:
-            # Simplified clustering
-            n_clusters = config.get('n_clusters', 3)
-            return np.random.randint(0, n_clusters, len(data))
-            
-        except Exception as e:
-            self.logger.error(f"Optimized clustering failed: {e}")
-            return np.zeros(len(data), dtype=int)
-    
-    def _perform_optimized_statistical_analysis(self, data: np.ndarray, config: Dict[str, Any]) -> Dict[str, Any]:
-        """Perform optimized statistical analysis."""
-        try:
-            # Simplified statistical analysis
-            return {
-                'mean': np.mean(data),
-                'std': np.std(data),
-                'correlation': np.corrcoef(data.T) if data.ndim > 1 else 1.0,
-                'statistical_significance': 0.95
-            }
-            
-        except Exception as e:
-            self.logger.error(f"Optimized statistical analysis failed: {e}")
-            return {'error': str(e)}
-    
-    def _optimize_regime_detection_results(self, result: Dict[str, Any]) -> Dict[str, Any]:
-        """Apply post-processing optimizations to regime detection results."""
-        try:
-            # Apply result optimizations
-            if 'regime_predictions' in result:
-                # Smooth predictions
-                result['regime_predictions'] = self._smooth_predictions(result['regime_predictions'])
-            
-            return result
-            
-        except Exception as e:
-            self.logger.warning(f"Result optimization failed: {e}")
-            return result
-    
-    def _smooth_predictions(self, predictions: np.ndarray) -> np.ndarray:
-        """Smooth regime predictions."""
-        try:
-            # Simple smoothing
-            smoothed = predictions.copy()
-            for i in range(1, len(smoothed) - 1):
-                if smoothed[i] != smoothed[i-1] and smoothed[i] != smoothed[i+1]:
-                    # Isolated prediction, smooth it
-                    smoothed[i] = smoothed[i-1]
-            
-            return smoothed
-            
-        except Exception as e:
-            self.logger.warning(f"Prediction smoothing failed: {e}")
-            return predictions
-    
-    def _track_performance(self, operation_type: str, execution_time: float, data_size: int):
-        """Track performance metrics."""
-        try:
-            # Record performance metrics
-            if self.performance_monitor:
-                metrics = self.performance_monitor['metrics']
-                metrics.execution_time = execution_time
-                metrics.throughput = data_size / execution_time if execution_time > 0 else 0.0
-                
-                # Update workload profiles
-                if operation_type not in self.workload_profiles:
-                    self.workload_profiles[operation_type] = {
-                        'total_time': 0.0,
-                        'total_operations': 0,
-                        'total_data_size': 0,
-                        'average_time': 0.0,
-                        'throughput': 0.0
-                    }
-                
-                profile = self.workload_profiles[operation_type]
-                profile['total_time'] += execution_time
-                profile['total_operations'] += 1
-                profile['total_data_size'] += data_size
-                profile['average_time'] = profile['total_time'] / profile['total_operations']
-                profile['throughput'] = profile['total_data_size'] / profile['total_time'] if profile['total_time'] > 0 else 0.0
-            
-        except Exception as e:
-            self.logger.error(f"Failed to track performance: {e}")
-    
-    def _fallback_regime_detection(self, data: np.ndarray, config: Dict[str, Any]) -> Dict[str, Any]:
-        """Fallback regime detection without optimization."""
-        n_regimes = config.get('n_regimes', 3)
-        return {
-            'regime_predictions': np.random.randint(0, n_regimes, len(data)),
-            'regime_probabilities': np.random.rand(len(data), n_regimes),
-            'regime_stability': np.random.rand(len(data)),
-            'transition_probabilities': np.random.rand(n_regimes, n_regimes)
-        }
-    
-    def _fallback_tree_operations(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Fallback tree operations without optimization."""
-        n_features = config.get('n_features', 10)
-        return np.random.rand(len(data), n_features)
-    
-    def _fallback_neural_inference(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Fallback neural inference without optimization."""
-        output_dim = config.get('output_dim', 1)
-        return np.random.rand(len(data), output_dim)
-    
-    def _fallback_clustering(self, data: np.ndarray, config: Dict[str, Any]) -> np.ndarray:
-        """Fallback clustering without optimization."""
-        n_clusters = config.get('n_clusters', 3)
-        return np.random.randint(0, n_clusters, len(data))
-    
-    def _fallback_statistical_analysis(self, data: np.ndarray, config: Dict[str, Any]) -> Dict[str, Any]:
-        """Fallback statistical analysis without optimization."""
-        return {
-            'mean': np.mean(data),
-            'std': np.std(data),
-            'correlation': 1.0,
-            'statistical_significance': 0.5
-        }
-    
-    def get_performance_summary(self) -> Dict[str, Any]:
-        """Get performance summary of the optimizer."""
-        try:
-            summary = {
-                'hardware_components': list(self.hardware_components.keys()),
-                'performance_monitor_enabled': self.performance_monitor is not None,
-                'memory_manager_enabled': self.memory_manager is not None,
-                'performance_history_count': len(self.performance_history),
-                'optimization_history_count': len(self.optimization_history),
-                'workload_profiles': self.workload_profiles,
-                'system_health': 'healthy'
-            }
-            
-            # Calculate overall performance metrics
-            if self.performance_history:
-                total_time = sum(h['execution_time'] for h in self.performance_history)
-                total_operations = len(self.performance_history)
-                total_data_size = sum(h.get('data_size', 0) for h in self.performance_history)
-                
-                summary['overall_performance'] = {
-                    'total_time': total_time,
-                    'total_operations': total_operations,
-                    'average_time': total_time / total_operations if total_operations > 0 else 0.0,
-                    'total_throughput': total_data_size / total_time if total_time > 0 else 0.0
-                }
-            
-            return summary
-            
-        except Exception as e:
-            self.logger.error(f"Failed to get performance summary: {e}")
-            return {'error': str(e)}
-    
-    def optimize_for_workload(self, workload_type: WorkloadType) -> Dict[str, Any]:
-        """Optimize system for specific workload type."""
-        try:
-            # Configure hardware for workload
-            self._configure_hardware_for_workload(workload_type)
-            
-            # Get optimization recommendations
-            recommendations = self._get_optimization_recommendations(workload_type)
-            
-            return {
-                'workload_type': workload_type.value,
-                'recommendations': recommendations,
-                'optimization_applied': True,
-                'timestamp': datetime.now().isoformat()
-            }
-            
-        except Exception as e:
-            self.logger.error(f"Failed to optimize for workload {workload_type}: {e}")
-            return {'error': str(e)}
-    
-    def _get_optimization_recommendations(self, workload_type: WorkloadType) -> List[str]:
-        """Get optimization recommendations for workload type."""
-        recommendations = []
-        
-        if workload_type == WorkloadType.REGIME_DETECTION:
-            recommendations.extend([
-                "Enable vectorization for regime detection",
-                "Use parallel processing for regime analysis",
-                "Optimize memory layout for regime data"
-            ])
-        elif workload_type == WorkloadType.TREE_OPERATIONS:
-            recommendations.extend([
-                "Enable parallel tree operations",
-                "Optimize tree traversal algorithms",
-                "Use efficient data structures"
-            ])
-        elif workload_type == WorkloadType.NEURAL_INFERENCE:
-            recommendations.extend([
-                "Enable GPU acceleration",
-                "Use mixed precision training",
-                "Optimize tensor operations"
-            ])
-        
-        return recommendations
-
-
-# Convenience functions
-def create_unified_hardware_optimizer(config: Optional[HardwareConfig] = None) -> UnifiedHardwareOptimizer:
-    """Create a unified hardware optimizer."""
-    if config is None:
-        config = HardwareConfig()
-    return UnifiedHardwareOptimizer(config)
-
-
-def quick_hardware_optimization(data: np.ndarray,
-                               workload_type: WorkloadType,
-                               config: Optional[HardwareConfig] = None) -> Dict[str, Any]:
-    """Quick hardware optimization with default settings."""
-    optimizer = create_unified_hardware_optimizer(config)
-    
-    if workload_type == WorkloadType.REGIME_DETECTION:
-        return optimizer.optimize_regime_detection(data, {'n_regimes': 3})
-    elif workload_type == WorkloadType.TREE_OPERATIONS:
-        return optimizer.optimize_tree_operations(data, {'n_features': 10})
-    elif workload_type == WorkloadType.NEURAL_INFERENCE:
-        return optimizer.optimize_neural_inference(data, {'output_dim': 1})
-    elif workload_type == WorkloadType.CLUSTERING:
-        return optimizer.optimize_clustering(data, {'n_clusters': 3})
-    elif workload_type == WorkloadType.STATISTICAL_ANALYSIS:
-        return optimizer.optimize_statistical_analysis(data, {})
-    else:
-        return {}
